requirements.txt
# Algorithmic Trading Backtester - Dependencies
# Generated: 2025-06-18

# Core Data Science & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Financial Data & Trading APIs
kiteconnect>=4.1.0          # Zerodha API
upstox-python-sdk>=2.0.0    # Upstox API (required for CLI usage)
yfinance>=0.2.18            # Yahoo Finance fallback
requests>=2.28.0            # HTTP client for APIs

# Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
Pillow>=10.0.0

# Configuration & Data Handling
PyYAML>=6.0
python-dotenv>=1.0.0

# Time & Date Utilities
pytz>=2023.3
python-dateutil>=2.8.0

# Performance & Parallel Processing
joblib>=1.3.0
multiprocessing-logging>=0.3.4

# Utilities
pathlib2>=2.3.7; python_version<"3.4"
typing-extensions>=4.5.0; python_version<"3.8"

# Development & Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: Advanced Analytics
scikit-learn>=1.3.0  # For ML-based strategies
statsmodels>=0.14.0  # For statistical analysis


# Optional: Database Support
SQLAlchemy>=2.0.0    # For data persistence
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Platform-specific dependencies
# Windows users may need: pywin32>=306
# macOS users may need: python-certifi-win32>=1.6
